requirements.txt
asgiref==3.8.1
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-jazzmin==3.0.0
django-tinymce==4.0.0
pillow==10.3.0
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
